Create TypingGame.js
This code creates a basic typing game with a list of words to type. When you type a word correctly and press Enter or Space, it moves to the next word. It calculates your typing speed in words per minute when you finish typing all the words in the list.